requirements.txt
fastapi==0.100.0
flake8==6.0.0
greenlet==2.0.2
isort==5.12.0
kaggle==1.5.15
numpy==1.25.1
pandas==2.0.3
pycodestyle==2.10.0
pydantic==2.0.2
pydantic_core==2.1.2
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.18
starlette==0.27.0
uvicorn==0.22.0
